Project
For this project I wanted to make a chat bot that people could talk to and vent all their problems and stress, I wanted to be fully interactive and make sure that the person using the bot is engaged, making sure people don't do certain stuff in accident, which is why they have to type something to start the bot, to make sure they are engaged and aware they started the bot. The reason why the code starts from end to the beginning is so everything is inside one code instead of having three different running functions for each part of the code, that way when the first function ends it automatically recalls the code behind it making it a smooth transition from code to code. The function vent_chat gathers the person's name and lets the user see what emotions can they vent about, it is just a friendly way to start the code, this code will then lead the person to the next function without having them to do anything. The function venting_emotion is for the person to tell me their mood based on the list, they also have the option to not talk to the bot in case they just wanted to say how they found themselves, in addition, if there was a typo or added other emotion they are reminded to only use what was in the list, this function will then lead the user to end_of_chat. End_of_chat is there for the purpose of making sure the user is enganged, that if they want to end chat is something they are making on purpose, but also as a way of still letting people vent in case they made a typo if they were too stressed, too anxious, too happy, too sad or just feeling too much to properly type.